Create MakeEnvelopes.swp
This .swp file, when opened in Solidworks 2018, will read a .txt file formatted as x y z \r\n and plot each z slice on a separate plane connected by lines. There is some basic lofting that works if the z slices are far enough apart, but this will be worked on in the future!